RavenDB-23560 refactor(ui): restructure sample queries into a shared reusable module and remove legacy components
