+++ src/Raven.Studio/typescript/components/common/sampleQueries/SampleQueriesPage.stories.tsx
