+++ src/Raven.Studio/typescript/components/common/sampleQueries/SampleQueriesPage.tsx
