+++ src/Raven.Studio/typescript/components/common/sampleQueries/partials/MethodsTable.tsx
