+++ src/Raven.Studio/typescript/components/common/sampleQueries/partials/SampleScriptsList.tsx
