+++ src/Raven.Studio/typescript/components/common/sampleQueries/partials/sampleQueriesTypes.ts
