+++ src/Raven.Studio/typescript/viewmodels/database/patch/PatchSampleQueriesPage.tsx
@@ -4,8 +4,8 @@ import { useAppSelector } from "components/store";
 import { databaseSelectors } from "components/common/shell/databaseSliceSelectors";
 import { Icon } from "components/common/Icon";
 import router from "plugins/router";
-import SampleQueriesPage from "./SampleQueriesPage";
-import { MethodGroup, SampleScript } from "./partials/sampleQueriesTypes";
+import SampleQueriesPage from "components/common/sampleQueries/SampleQueriesPage";
+import { MethodGroup, SampleScript } from "components/common/sampleQueries/partials/sampleQueriesTypes";
 import AboutViewFloating, { AccordionItemWrapper } from "components/common/AboutView";
 import savedPatchesStorage from "common/storage/savedPatchesStorage";
 import messagePublisher from "common/messagePublisher";
